requirements.txt
asgiref==3.5.2
attr==0.3.2
attrs==20.3.0
blinker==1.5
ConfigParser==5.3.0
cryptography==38.0.3
Cython==0.29.32
dl==0.1.0
docutils==0.19
drawnow==0.72.5
Flask_Login==0.6.2
Flask_WTF==1.0.1
HTMLParser==0.0.2
ipython==8.6.0
ipywidgets==8.0.2
jnius==1.1.0
keyring==23.11.0
lockfile==0.12.2
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
protobuf==4.21.9
pyOpenSSL==22.1.0
pyserial==3.5
python-dotenv==0.21.0
railroad==0.5.0
scipy==1.9.3
Sphinx==5.3.0
tornado==6.2
trove_classifiers==2022.10.19
truststore==0.5.0
urllib3_secure_extra==0.1.0
watchdog==2.1.9
WTForms==3.0.1
xmlrpclib==1.0.1
